Revert "[refactor] move truncate to deletion operator"
This reverts commit 05f9f2f145e338ab4e01566eec4f1a49331c128d.
