+++ pkg/sql/compile/compile.go
@@ -1133,16 +1133,20 @@ func (c *Compile) compilePlanScope(ctx context.Context, step int32, curNodeIdx i
 	case plan.Node_DELETE:
 		c.appendMetaTables(n.DeleteCtx.Ref)
 		if n.DeleteCtx.CanTruncate {
-			rs := c.newMergeScope(nil)
 			arg, err := constructDeletion(n, c.e, c.proc)
 			if err != nil {
 				return nil, err
 			}
-			rs.Instructions = append(rs.Instructions, vm.Instruction{
-				Op:  vm.Deletion,
-				Arg: arg,
-			})
-			ss := []*Scope{rs}
+			ss := []*Scope{{
+				Magic: Deletion,
+				Plan:  c.pn,
+				Instructions: vm.Instructions{
+					{
+						Op:  vm.Deletion,
+						Arg: arg,
+					},
+				},
+			}}
 			return ss, nil
 		}
 		curr := c.anal.curr
